refactor(core): extract bounce and canary persistence from ledger.ts
Move recordBounce, suppressionFor, contactSuppressionFor,
bounceStatsByIdentity, listRecentBounces, countBounces,
countAutoPermanentBounces, recordCanaryResult, latestCanaryResult, and
latestSentEmailCopy out of Ledger's inline SQL into a new
packages/core/src/delivery-health.ts module: pure functions of a raw
Database handle, mirroring the ledger-schema.ts pattern from #452.

Ledger's methods now delegate to the new module with identical
signatures and SQL, so the exported Ledger API, suppression semantics,
timestamps, transactions, and every call site are unchanged.

Records the completed slice in ROADMAP.md's ledger-split tech-debt item
(Approved-not-yet-started tracking line left untouched, per policy).

Fixes #617
